added hashing and basic routers
