+++ app/main.py
@@ -3,12 +3,12 @@ Module Docstring
 """
 
 import uvicorn
-from fastapi import Depends, FastAPI, HTTPException, status
-from sqlalchemy.orm import Session
+from fastapi import FastAPI
 
-from db import models, schemas
+from app.routers import posts, users
+from db import schemas
 from db.db_orm import Base, database_gen, engine
-from db.db_utils import init_db, stream_mocker
+from db.db_utils import init_db  # , stream_mocker
 from db.repository import PostgresDB
 from grafana.grafana_utils import Grafana
 
@@ -26,12 +26,13 @@ grafana.add_database_source(postgres_db)
 
 
 #####################      Simulating a stream of posts    ##################################
-stream_mocker(get_initial_db)
+# stream_mocker(get_initial_db)
 
 ##############################    Creatng FastAPI App   ##########################
 app = FastAPI()
 
-print()
+app.include_router(posts.router)
+app.include_router(users.router)
 
 
 @app.get("/")
@@ -40,125 +41,6 @@ async def root():
     function docstring
     """
     return {"message": "Hello World 2"}
-
-
-@app.get("/posts", response_model=list[models.PostResponse])
-def get_all_posts(db_session: Session = Depends(database_gen)):
-    """
-    function docstring
-    """
-    posts = db_session.query(schemas.Post).all()
-    # .execute("SELECT * FROM posts")
-    return posts
-
-
-@app.post("/posts", status_code=status.HTTP_201_CREATED, response_model=models.PostResponse)
-def create_post(payload: models.PostCreate, db_session: Session = Depends(database_gen)):
-    """
-    function docstring
-    """
-    # creating a data according to schema
-    new_post = schemas.Post(**(payload.dict()))
-    # adding data to session, moving it to pending state.
-    db_session.add(new_post)
-    # moving all data in pending state, in this session, to persistant state.
-    db_session.commit()
-    # update new_post with data returned from db_session
-    db_session.refresh(new_post)
-    return new_post
-
-
-# path operations are evaluated in order,
-# you need to make sure that the path for /posts/latest
-# is declared before the one for /posts/{identifier}
-@app.get("/posts/latest", response_model=models.PostResponse)
-def get_latest_post(db_session: Session = Depends(database_gen)):
-    """
-    function docstring
-    """
-    length_db = db_session.query(schemas.Post).count()
-    post_wanted = db_session.query(schemas.Post).where(schemas.Post.id == length_db).first()
-    return post_wanted
-
-
-# identifier is an example of a path parameter
-@app.get("/posts/{identifier}", response_model=models.PostResponse)
-def get_post(identifier: int, db_session: Session = Depends(database_gen)):
-    """
-    Creates endpoint to fetch specific post
-    """
-    post_wanted = db_session.query(schemas.Post).where(schemas.Post.id == identifier).first()
-    if post_wanted is None:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail=f"Post with id {identifier} not found!",
-        )
-    # response.status_code = status.HTTP_404_NOT_FOUND
-    return post_wanted
-
-
-# here identifier is a Query parameter
-@app.delete("/posts", status_code=status.HTTP_204_NO_CONTENT)
-def delete_post(identifier: int, db_session: Session = Depends(database_gen)):
-    """
-    Function that creates the resource to delete a specified post, by identifier.
-    """
-    post_wanted = db_session.query(schemas.Post).where(schemas.Post.id == identifier).first()
-    if post_wanted is None:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail=f"Post with id {identifier} not found!",
-        )
-    db_session.delete(post_wanted)
-    db_session.commit()
-
-
-@app.patch(
-    "/posts/{identifier}", status_code=status.HTTP_200_OK, response_model=models.PostResponse
-)
-def patch_post(
-    identifier: int, payload: models.PostUpdate, db_session: Session = Depends(database_gen)
-):
-    """
-    function docstring
-    """
-    post_wanted = db_session.query(schemas.Post).where(schemas.Post.id == identifier)
-    if post_wanted.first() is None:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail=f"Post with id {identifier} not found!",
-        )
-    # pylance type checker says that payload.dict() is incompatible with type of `values` from
-    # update... hence the extra dict()
-    post_wanted.update(dict(payload.dict()), synchronize_session=False)
-    db_session.commit()
-    return post_wanted.first()
-
-
-#################################        User Endpoints        ################################
-@app.get("/users", response_model=list[models.UserResponse])
-def get_all_users(db_session: Session = Depends(database_gen)):
-    """
-    function docstring
-    """
-    users = db_session.query(schemas.User).all()
-    return users
-
-
-@app.post("/users", status_code=status.HTTP_201_CREATED, response_model=models.UserResponse)
-def create_user(payload: models.UserCreate, db_session: Session = Depends(database_gen)):
-    """
-    function docstring
-    """
-    # creating a data according to schema
-    new_user = schemas.User(**(payload.dict()))
-    # adding data to session, moving it to pending state.
-    db_session.add(new_user)
-    # moving all data in pending state, in this session, to persistant state.
-    db_session.commit()
-    # update new_post with data returned from db_session
-    db_session.refresh(new_user)
-    return new_user
 
 
 if __name__ == "__main__":
